% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BubbleEnrichShiny.R
\name{BubbleEnrichShiny}
\alias{BubbleEnrichShiny}
\title{shiny app for BubbleEnrich Package}
\usage{
BubbleEnrichShiny()
}
\value{
No return value but open up a shiny page.
}
\description{
A function that launches the shiny app for this package.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
BubbleEnrichShiny()
}

}
