% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BubbleGSEA.R
\name{BubbleGSEA}
\alias{BubbleGSEA}
\title{\code{BubbleGSEA} gets a gene set of interest and outputs a bubble graph with bubbles diameters
correlated to the significance of the disease phenotype to the gene set and a tree diagram spreading
from the bubbles of the specific genes that contribute to the phenotypic output.}
\usage{
BubbleGSEA(geneList)
}
\arguments{
\item{geneList}{- the gene set of interest should be as a list that satisfies the 3 features:
1.numeric vector: fold change or other type of numerical variable
2.named vector: every number has its corresponding gene ID
3.sorted vector: should be sorted in decreasing order}
}
\value{
(data.frame) - a bubble plot of the enrichment results of the gene set and the dataframe that the bubble plot is based on
}
\description{
\code{BubbleGSEA} gets a gene set of interest and outputs a bubble graph with bubbles diameters
correlated to the significance of the disease phenotype to the gene set and a tree diagram spreading
from the bubbles of the specific genes that contribute to the phenotypic output.
}
\examples{
\dontrun{
library(DOSE)
data(geneList)
#uses the example dataset of breast cancer from the DOSE package
BubbleGSEA(geneList)
}
}
\seealso{
TODO : Documentation of semantic similarity calculations and GO enrichment analysis \code{\link{GOSemSim}}
}
\author{
\href{https://orcid.org/0000-0003-4609-4965}{Cathy Cha} (aut)
}
