% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourInput.R
\name{colourInput}
\alias{colourInput}
\title{Helper Function for BubbleEnrich shiny application}
\usage{
colourInput(colour, type)
}
\arguments{
\item{colour}{A string corresponding to userinput colour selection}

\item{type}{A string representing the value of the colour (low or high p-value)}
}
\value{
A string that represents a colour in ggplot
}
\description{
A function that returns a colour scheme based on user input in the BubbleEnrich
shiny applicaton
}
\examples{
colourInput("Purple", "highcolour")
colourInput("Blue", "lowcolour")

}
\author{
\href{https://orcid.org/0000-0003-4609-4965}{Cathy Cha} (aut)
}
