% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disgeneAnnot}
\alias{disgeneAnnot}
\title{Disease to Gene Annotations from DisGeNET}
\format{A dataframe with 1548061 rows and 3 columns
\describe{
 \item{geneID}{NCBI Entrez Gene Identifier}
 \item{diseaseId}{UMLS concept unique identifier}
 \item{diseaseName}{Name of the disease}
}}
\source{
\url{http://www.disgenet.org/static/disgenet_ap1/files/downloads/all_gene_disease_pmid_associations.tsv.gz}
}
\usage{
disgeneAnnot
}
\description{
Disease to Gene Annotations from DisGeNET
}
\examples{
\dontrun{
 disgeneAnnot
}
}
\keyword{datasets}
