% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userInput.R
\name{userInput}
\alias{userInput}
\title{\code{userInput}}
\usage{
userInput(geneSet)
}
\arguments{
\item{geneSet}{- string path to the gene set user wants to use, store in a csv file}
}
\value{
a numeric vector sorted in decreasing order with name as geneID and numeric value as fold change
}
\description{
\code{userInput}
}
\examples{
\dontrun{

numVec <- userInput("")
}

}
\author{
\href{https://orcid.org/0000-0003-4609-4965}{Cathy Cha} (aut)
}
