% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BubbleEnrich.R
\name{BubbleEnrich}
\alias{BubbleEnrich}
\title{Bubble Graph of Disease Phentype Enrichment Results}
\usage{
BubbleEnrich(geneList = NULL)
}
\arguments{
\item{geneList}{A numeric, named vector, sorted in descending order}
}
\value{
A ggplot plot
}
\description{
This function gets a ranked gene set of interest with related phenotypes and outputs a
bubble graph with bubble diameters correlated to the significance of the phenotype to the
gene set.
}
\examples{
\dontrun{
data(geneList, package="DOSE")
BubbleEnrich(geneList)
}
}
\references{
Guangchuang Yu, Li-Gen Wang, Yanyan Han and Qing-Yu He. clusterProfiler: an R package
for comparing biological themes among gene clusters. OMICS: A Journal of Integrative
Biology 2012, 16(5):284-287 \href{https://www.ncbi.nlm.nih.gov/pubmed/22455463}
}
\author{
\href{https://orcid.org/0000-0003-4609-4965}{Cathy Cha} (aut)
}
