% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BubbleEnrich.R
\name{BubbleEnrich}
\alias{BubbleEnrich}
\title{\code{BubbleEnrich} gets a ranked gene set of interest with related phenotypes and outputs a
bubble graph with bubbles diameters correlated to the significance of the phenotype to the
gene set and corresponding genes contributing to the enrichment spreading from the disease bubbles.}
\usage{
BubbleEnrich(geneList)
}
\arguments{
\item{geneList}{- a numeric, named vector, sorted in descending order}
}
\description{
\code{BubbleEnrich} gets a ranked gene set of interest with related phenotypes and outputs a
bubble graph with bubbles diameters correlated to the significance of the phenotype to the
gene set and corresponding genes contributing to the enrichment spreading from the disease bubbles.
}
\examples{
\dontrun{
data(geneList, package="DOSE")
BubbleEnrich(geneList)
}
}
\author{
\href{https://orcid.org/0000-0003-4609-4965}{Cathy Cha} (aut)
}
