% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disgeneAnnot}
\alias{disgeneAnnot}
\title{Disease to Gene Annotations from DisGeNET}
\format{A dataframe with 1548061 rows and 15 columns
\describe{
 \item{geneID}{hello}
 \item{geneSymbol}{}
 \item{DSI}{}
 \item{DPI}{}
 \item{diseaseId}{}
 \item{diseaseName}{}
 \item{diseaseType}{}
 \item{diseaseSemanticType}{}
 \item{score}{}
 \item{EI}{}
 \item{YearInitial}{}
 \item{YearFinal}{}
 \item{pmid}{}
 \item{source}{}
}}
\source{
\url{http://www.disgenet.org/static/disgenet_ap1/files/downloads/all_gene_disease_pmid_associations.tsv.gz}
}
\usage{
disgeneAnnot
}
\description{
Disease to Gene Annotations from DisGeNET
}
\examples{
\dontrun{
 disgeneAnnot
}
}
\keyword{datasets}
