% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneIDsort.R
\name{geneIDsort}
\alias{geneIDsort}
\title{\code{userInput}}
\usage{
geneIDsort(enrichData)
}
\arguments{
\item{enrichData}{- dataframe of the disease phenotypes with its corresponding geneIDs that contribute to it's enrichment}
}
\value{
a vector object with phenotype with it's corresponding geneids
}
\description{
\code{userInput}
}
\examples{
\dontrun{

}

}
\author{
\href{https://orcid.org/0000-0003-4609-4965}{Cathy Cha} (aut)
}
