% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HGNCAnnot}
\alias{HGNCAnnot}
\title{geneId to HGNC approved symbol from HGNC}
\format{A dataframe with 46736 rows and 2 columns
\describe{
 \item{HGNC.ID}{HGNC gene ID}
 \item{Approved.symbol}{HGNC approved gene symbols}
}}
\usage{
HGNCAnnot
}
\description{
geneId to HGNC approved symbol from HGNC
}
\examples{
\dontrun{
 HGNCAnnot
}
}
\keyword{datasets}
