% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userInput.R
\name{userInput}
\alias{userInput}
\title{Helper Function for BubbleEnrich shiny application}
\usage{
userInput(geneSet)
}
\arguments{
\item{geneSet}{A string path to the gene set user wants to use, stored in a csv file}
}
\value{
a numeric vector sorted in decreasing order with name as geneID and numeric value as fold change
}
\description{
A function that formats a .csv user input of gene set data into a numerical vector in
decreasing order.
}
\examples{
\dontrun{
numVec <- userInput("/extdata/data.csv")
}

}
\author{
\href{https://orcid.org/0000-0003-4609-4965}{Cathy Cha} (aut)
}
